+++ src/features/actions/List.svelte
@@ -2,7 +2,7 @@
   import WorkflowRun from './WorkflowRun.svelte';
   import githubSVG from '$assets/github-logo.svg';
   import { repositoryCollapseStore } from '$shared/stores/repository-collapse.store';
-  import CountBadge from '$shared/components/CountBadge.svelte';
+  import CountBadge from '$shared/ui/CountBadge.svelte';
 
   let { org, repo, workflowRuns = [] } = $props();
