+++ src/shared/ui/CountBadge.svelte
